feat(OUT-1400) : transferred comment create notification to a trigger job
