+++ package.json
@@ -69,6 +69,7 @@
     "supabase": "^2.1.1",
     "swr": "^2.2.5",
     "tapwrite": "1.1.82",
+    "trigger.dev": "3.3.13",
     "zod": "^3.23.8"
   },
   "devDependencies": {
