+++ src/jobs/notifications/index.ts
@@ -1,3 +1,4 @@
 export { deleteTaskNotifications } from './delete-task-notifications'
 export { sendTaskCreateNotifications } from './send-task-create-notifications'
 export { sendTaskUpdateNotifications } from './send-task-update-notifications'
+export { sendCommentCreateNotifications } from './send-comment-create-notifications'
